refactor(core/type): [Wrapper] rename to types; remove namespaces
