+++ packages/next-starter/src/modules/core/redux/interfaces/index.ts
@@ -1,3 +1,3 @@
 export * from './Action.d';
 export * from './Store.d';
-export * from './Wrapper.d';
+export * from './types.d';
